Disabled currency conversion tests
That function is no longer what we’re actually using and we don’t actually use the convert endpoint. The tests take forever because that function seems much slower than our client side one, and it seems to take longer as we add more sources. Will reenable once we get the up to date convert function in this repo.